clean up + start of css export
