+++ bin/export_svg/astro-icon-export.sketchplugin/Contents/Sketch/manifest.json
@@ -1,9 +1,9 @@
 {
-  "author" : "Aleksandr Antonov",
+  "author" : "Rocket Communications",
   "commands" : [
     {
       "script" : "script.js",
-      "name" : "export-svg",
+      "name" : "astro-icon-export",
       "handlers" : {
         "run" : "onRun"
       },
@@ -11,7 +11,7 @@
     }
   ],
   "menu" : {
-    "title" : "export-svg",
+    "title" : "astro-icon-export",
     "items" : [
       "com.bohemiancoding.sketch.runscriptidentifier"
     ]
@@ -20,5 +20,5 @@
   "version" : "1.0",
   "description" : "",
   "authorEmail" : "",
-  "name" : "export-svg"
+  "name" : "astro-icon-export"
 }
