+++ package.json
@@ -15,6 +15,7 @@
     "url": "https://github.com/RocketCommunicationsInc/astro-components/src/master"
   },
   "description": "Astro Web Component for buttons",
+  "main": "bin/index.js",
   "license": "SEE LICENSE IN LICENSE.md",
   "private": false,
   "scripts": {
@@ -30,7 +31,8 @@
     "lit-html": "^1.2.1",
     "military-timezones": "^1.0.0",
     "prompt-sync": "^4.2.0",
-    "svgo": "^1.3.2"
+    "svgo": "^1.3.2",
+    "yargs": "^16.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
