+++ src/css/components/icons-detail.css
@@ -1,0 +1,3 @@
+.rux-icon-vibration{
+    background: url("/icons/astro-new.svg#vibration") no-repeat;
+}
